backend/requirements.txt
# Flask core
Flask==3.1.0
flask-cors==5.0.1
flask-limiter==3.12

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
alembic==1.14.1
Flask-Migrate==4.1.0

# Authentication
flask-jwt-extended==4.7.1
bcrypt==4.2.1

# Environment
python-dotenv==1.0.1

# PDF Generation
WeasyPrint==63.1

# Google Gemini AI
google-genai==1.14.0

# Validation
marshmallow==3.23.2

# Testing
pytest==8.3.4
pytest-flask==1.3.0
